requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-html==4.1.1
click==8.1.7
jinja2==3.1.2
streamlit==1.32.0